+++ js/main.js
@@ -1,7 +1,6 @@
 import { getPhotosDataArray } from './data.js';
-import { renderThumbnails } from './thumbnails-render.js';
+import { renderThumbnails } from './thumbnails.js';
 import './photo-modal.js';
 
 getPhotosDataArray();
 renderThumbnails();
-
